requirements.txt
requests
psycopg2-binary
pydantic-settings
pytest
